+++ NEWS.md
@@ -13,6 +13,7 @@ Add the link to [Aegis Privacy Shield & Terminator](https://github.com/yarrowmar
 
 New presets/rules:
 
+* 2026/08/18 - `DisableWERCPlSupport`/`EnableWERCPlSupport`/`ViewWERCPlSupport` -> Disable Windows Error Reporting Control Panel Support
 * 2026/08/18 - `DisableCDPSvc`/`EnableCDPSvc`/`ViewCDPSvc` -> Disable Connected Devices Platform Service
 * 2026/08/18 - `DisablePcaSvc`/`EnablePcaSvc`/`ViewPcaSvc` -> Disable Windows Program Compatibility Assistant Service
 * 2026/07/17 - `DisableThirdPartySuggestions_CU`/`EnableThirdPartySuggestions_CU`/`ViewThirdPartySuggestions_CU` -> Disable Third Party Suggestions